move jena models
