+++ integration/src/test/scala/org/knora/webapi/messages/util/rdf/JsonLDUtilSpec.scala
@@ -12,7 +12,7 @@ import java.nio.file.Paths
 
 import org.knora.webapi.CoreSpec
 import org.knora.webapi.messages.util.rdf._
-import org.knora.webapi.messages.util.rdf.jenaimpl.JenaModelFactory
+import org.knora.webapi.messages.util.rdf.JenaModelFactory
 import org.knora.webapi.util.FileUtil
 
 /**
