+++ integration/src/test/scala/org/knora/webapi/store/triplestore/upgrade/plugins/UpgradePluginPR1322Spec.scala
@@ -6,7 +6,7 @@
 package org.knora.webapi.store.triplestore.upgrade.plugins
 
 import org.knora.webapi.messages.util.rdf._
-import org.knora.webapi.messages.util.rdf.jenaimpl.JenaRepository
+import org.knora.webapi.messages.util.rdf.JenaRepository
 
 class UpgradePluginPR1322Spec extends UpgradePluginSpec {
   "Upgrade plugin PR1322" should {
