+++ webapi/src/main/scala/org/knora/webapi/messages/util/rdf/RdfFeatureFactory.scala
@@ -5,8 +5,6 @@
 
 package org.knora.webapi.messages.util.rdf
 
-import org.knora.webapi.messages.util.rdf.jenaimpl.*
-
 /**
  * A feature factory that creates RDF processing tools.
  */
